Gaming Hub API/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.2
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
Pillow==10.1.0
passlib==1.7.4
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.0.1
python-multipart==0.0.22
redis==7.2.1
requests==2.31.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
